pyproject.toml
[tool.poetry]
name = "spakky-fastapi"
version = "2.0.0"
description = "Highly abstracted Framework core to use DDD & DI/IoC & AOP & Etc..."
authors = ["Spakky <sejong418@icloud.com>"]
readme = "README.md"

[tool.poetry.plugins."spakky.plugins"]
fastapi = "spakky_fastapi.main:initialize"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.115.5"
spakky-core = ">=2.10"
websockets = "^14.1"
orjson = "^3.10.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-xdist = "^3.6.1"
toml = "^0.10.2"
httpx = "^0.27.2"
ruff = "^0.9.7"
pre-commit = "^4.1.0"

[tool.poetry.group.ci.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-xdist = "^3.6.1"
httpx = "^0.27.2"
ruff = "^0.9.7"

[tool.ruff]
builtins = ["_"]
cache-dir = "~/.cache/ruff"

[tool.pytest.ini_options]
pythonpath = "spakky"
testpaths = "tests"
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = """
    --cov
    --cov-report=term
    --cov-report=xml
    --no-cov-on-fail
    --strict-markers
    --dist=loadfile
    -p no:warnings
    -n auto
    -vv
"""

[tool.coverage.run]
omit = ["tests/*"]
traceback = "short"
branch = true

[tool.coverage.report]
show_missing = true
precision = 2
fail_under = 90
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "@(typing\\.)?overload",
]
